refactor: split constraint generation operations
