+++ fm_generator/src/fm_generator/FMGenerator/operations/generate_models.py
@@ -15,6 +15,13 @@ from fm_generator.FMGenerator.operations.satisfiability import (
     is_model_satisfiable,
 )
 
+__all__ = [
+    "SATISFIABILITY_MAX_ATTEMPTS",
+    "SAT_SEED_STRIDE",
+    "generate_single_model",
+    "is_model_satisfiable",
+]
+
 
 def build_uvl_includes(params: Params) -> list[str]:
     includes: list[str] = []
